gro: smart search refactor
